+++ src/modules/middlewares/headers.interceptor.ts
@@ -22,7 +22,7 @@ import { GENERAL_HEADERS } from 'utils/constants.ts'
  * - `x-znx-<type>-session-status:<SessionStatus}>` is added to indicate the session status.
  * - `x-znx-<type>-id` Subject Id header is added when a user token identifier (`sub`) is included.
  * - If `X-Znx-Cookies-Accepted: true` is present, session cookies are sent in the `Set-Cookie` header:
- *   - `x-znx-<type>-session-status=<SessionStatus>; x-znx-<type>-id=<sub>; Max-Age=<seconds>; Path=/; HttpOnly; SameSite=Strict`
+ *   - `x-znx-app-token=<sessionToken>; x-znx-<type>-session-status=<SessionStatus>; x-znx-<type>-id=<sub>; Max-Age=<seconds>; Path=/; HttpOnly; SameSite=Strict`
  * - `Max-Age` is calculated from the session expiration timestamp minus the current Unix time.
  *
  * @returns {MiddlewareInterceptor}
@@ -32,24 +32,28 @@ import { GENERAL_HEADERS } from 'utils/constants.ts'
 export const sessionHeadersInterceptor = (): MiddlewareInterceptor => {
   const { cookiesAcceptedHeader } = GENERAL_HEADERS
 
-  return ({ locals: { session }, req: { headers } }, response) => {
+  return (ctx, response) => {
+    const { locals: { session }, req: { headers }, cookies } = ctx
     if (!session?.type) return response
 
     const cookiesAccepted = headers.get(cookiesAcceptedHeader) === 'true'
-    const { payload, type, subject, status } = session
+    const { payload, type, subject, status, token } = session
     const authSessionType = type === 'anonymous' ? 'user' : type
 
     const sessionHeaders = getSessionHeaders({
-      subject: subject || getClientSubject(headers, authSessionType) || session.id,
+      subject: subject || getClientSubject(headers, cookies, authSessionType) || session.id,
       expiration: payload?.exp,
       sessionStatus: status,
       type: authSessionType,
+      refreshToken: token,
       cookiesAccepted,
     })
 
     for (const header of Object.entries(sessionHeaders)) {
       response.headers.append(...header)
     }
+
+    delete ctx.locals.session
 
     return response
   }